refactor(render): Move `generate_content` into `render`
