+++ adr_viewer/__init__.py
@@ -1,25 +1,10 @@
-import os
 from typing import List
 
 from click import option, command
 
 from adr_viewer.parse import parse_adr, parse_adr_files, Adr
-from adr_viewer.render import render_html, AdrTemplateConfig
+from adr_viewer.render import render_html, AdrTemplateConfig, generate_content
 from adr_viewer.server import run_server
-
-
-def generate_content(adrs: List[Adr], template_dir_override=None, title=None) -> str:
-    config = AdrTemplateConfig(
-        project_title=title if title else os.path.basename(os.getcwd()), records=[]
-    )
-
-    for index, adr in enumerate(adrs):
-        adr.index = index
-        adr.includes_mermaid |= config.include_mermaid
-
-        config.records.append(adr)
-
-    return render_html(config, template_dir_override)
 
 
 # fmt: off
