+++ adr_viewer/render.py
@@ -1,3 +1,4 @@
+import os
 from dataclasses import dataclass
 from typing import List
 
@@ -30,3 +31,17 @@ def render_html(config: AdrTemplateConfig, template_dir_override=None) -> str:
     template = env.get_template("index.html")
 
     return template.render(config=config)
+
+
+def generate_content(adrs: List[Adr], template_dir_override=None, title=None) -> str:
+    config = AdrTemplateConfig(
+        project_title=title if title else os.path.basename(os.getcwd()), records=[]
+    )
+
+    for index, adr in enumerate(adrs):
+        adr.index = index
+        adr.includes_mermaid |= config.include_mermaid
+
+        config.records.append(adr)
+
+    return render_html(config, template_dir_override)
